backend/requirements.txt
# Core
fastapi>=0.95.0
uvicorn>=0.21.0
python-dotenv>=1.0.0
httpx>=0.23.0

# Telegram Bot
python-telegram-bot>=20.0

# Minimal dependencies for basic functionality
numpy>=1.24.0
requests>=2.28.0